Allow the time-series name to be retrieved from the result row
InfluxDB supports querying multiple time-series at the same time.

With that in mind, it's crucially important to figure out which
result row comes from which time-series in order to make sense
of the data.

The current implementation preserves backwards compatibility
with the old way of doing things with the only exception of
throwing an InvalidArgumentException exception when accessing
a missing key.